requirements.txt
streamlit>=1.37.0
cryptography>=43.0.1
requests>=2.32.4
python-dotenv
pydantic==2.5.0
python-multipart>=0.0.18
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-jose[cryptography]>=3.4.0
passlib[bcrypt]==1.7.4
playwright>=1.40.0
pandas==2.3.3
x402[evm,svm]>=2.0.0
httpx>=0.27.0
eth-account>=0.13.0